physics: Unit dyads access from ReferenceFrames
Unit dyads can be accessed through attributes of a ReferenceFrame
object. It is an addition related to issue #24965, allowing easy access
to unit dyads without the need to use the outer() function. This commit
adds 9 functions with the @property decorator, along with a couple of
examples in the constructor. Access to unit dyadic is not implemented.