physics.continuum_mechanics: Fixed the ramp load and added Doctest in Beam
Added a dividing factor for ramp loads which was needed to be divided to get the right answer.
The dividing factor is equal to (end-start) for order 1 (ramp loads).
It has been added to the functions apply_load(), remove_load(), _handle_end().
The constant distributed load and parabolic ramp loads are satisfying the doctests, Hence,
dividing_factor is set to 1. Previously, no Doctest or Test was added to test the ramp load.
So, this error had been overlooked. Added a Doctest with ramp load to validate the changes.

Examples:
There is a beam of length 6 meters. A ramp load of magnitude 4 N/m is
applied in the downward direction starting from 3 meters away from the
starting point of the beam till the end of the beam. Find the expression
for the load.

>>> from sympy.physics.continuum_mechanics.beam import Beam
>>> from sympy import symbols
>>> E, I = symbols('E, I')
>>> b = Beam(6, E, I)
>>> b.apply_load(4, 3, 1, end=6)
>>> b.load

Result:
4*SingularityFunction(x, 3, 1)/3 - 4*SingularityFunction(x, 6, 0) - 4*SingularityFunction(x, 6, 1)/3

Which is the correct answer.
We can verify this by seeing that the value of load at x=3 is 0 and at x=6 is 4.