[physics.continuum_mechanics] find the maximum positive and maximum negative shear forces and bending moments
In many problems we would not just want to know the maximum value of
the shear force and bending moment but also the max negative or max
positive values for the same. Then, the absolute maximum is being calculated
using this. the max negative or max positive values could be retrieved by either
making it a property of the beam object or by printing it in the function.