integrals: Fixed bug with sqrt(a) factors in Erf/Fresnel rules
Rules in `manualintegrate.py` changed to use `sqrt(pi)/sqrt(a)` factors
instead of `sqrt(pi/a)` so that cases where `a` is negative are handled
correctly. Tests for case when `a` not positive added and one test in
`test_integrals.py` revised to pass with new formulation.

Addresses issue https://github.com/sympy/sympy/issues/25093