Simplified hierarchy of Plot class and *Backend classes
The `Plot` class used OOP composition with respect to a `*Backend` class.
This PR changes it to inheritance: the `Plot` class becomes the base class
for `*Backend` classes, holding the most common attributes exptected to be used
by most plotting libraries, like title, axis labels, etc.
In doing so, the `BaseBackend` and `DefaultBackend` classes are no longer needed:
their logic have been incorporated into the `Plot` class.

Previously, `MatplotlibBackend` had to keep track of all the subplot-axes
required by the `PlotGrid` class. For example, a PlotGrid containing 4 plots
created instantiated a single `MatplotlibBackend` which create a figure with
4 subplots` (hence, the `ax` attribute was a list of matplotlib's axes).
This PR changes it: `MatplotlibBackend` is only going to deal with one axes
(one subplot). As a consequence, `PlotGrid` was modified to create the overall
figure and subplot-axes, which are going to be fed to the backend.

This simplifies the overall structure, while allowing new behaviors to be
implemented on future PR, specifically:

1. interactive widget plots.
2. more complex grid layouts with PlotGrid.