Fixes the satask portion of the bug from #25349
Before this fix, extract_predargs was not extracting arguements
that were only present in the assumptions in some cases.
As a result, satask would fail to raise an inconsistent assumptions
exception in some cases. This is now fixed.

An unfortunate consequence of this fix is that it makes satask
slower. Previously running the tests for the new assumptions
took 21.7 +/- .2 seconds for me. Now they take 26.5 +/- .2
seconds.