Improved `ecm` stage2
As pointed out in [1], the prime numbers used in stage2 can be computed
in advance. In particular, the x-coordinates of P and -P are identical,
thus reducing the number of calculations.
For example, since 50+-3 are both prime numbers, there is no need to check
47 and 53 respectively, but only one of them. This means that although there
are 24 primes from 3 to 100, only 18 checks (centered on 50) are needed.

In addition, the following corrections were made:

* Added docstring description.
* Changed from `integer_log` to `math.log`. At least, `math.log` is faster
  in the practical range.
* Fixed a problem where `T` was not set properly when `sqrt(B2)` was too small.
* Checking if `B1` or `B2` is even is now done by `ecm`.

[1] Carl Pomerance, Richard Crandall, Prime Numbers:
    A Computational Perspective, 2nd Edition (2005), ISBN:978-0387252827