Started working on the Transformer.
Folded a bunch of nodes in the Lark grammar to simplify the Transformer class.

Revamped the `temporary_helper.py` file to emit more useful stats while working on the transformer.

Updated the old transformer code I had written (for experimenting purposes) for the new grammar, and removed a bunch of code for the now-useless nodes.