Introducing get_data for plotting series
The plotting module implements `*Series` classes in order to generate numerical data
for plots. Each class exposes a different method to retrieve the numerical data.
For example, line-related series expose `get_points`, surface-related series expose
`get_meshes`, implicit series exposes `get_raster`.

As the number of `*Series` classes is going to increase, a common method to retrieve
numerical data is introduced: `get_data`. This also allows to return a different
number of elements in comparison to the aforementioned methods, which in turns allow
to extend the capabilities of the plotting module.