Updated the grammar.
Removed a bunch of unnecessary imports, modified the definition of the `number` node, and removed a bunch of unnecessary comments.

Modified the definition of `sub` to have two specific cases.

Folded `fraction` so that it doesn't show up in the parse tree.

Broke up the definition of `fraction` and `binomial` into 3 separate cases.