Added more features to the transformer.
Added support for the `DIGIT` node.

Modified and finalized the transformer code for the `number` node.

Added support for fractions.

Updated a comment in the `integral` node.

Added support for the `group_curly_parentheses_special` node, which is required for the
`summation` and `product` nodes.

Updated a comment in the `remove_tokens` function.

Added a couple of short forms for commonly-used debug and testing functions.