Fixed some bugs in the `integral` code for the transformer.
Removed a ill-formed integral-related test, and updated some integral-related tests.

Updated a sum-related test.

Added some expected failures for the integral-related tests.

Added the `XFAIL` marker to the derivative tests since that feature is yet to be added.

Added the `XFAIL` marker to the miscellaneous tests since those expressions are suspect.

Added the `XFAIL` marker to the spacing-command-related tests since there is an unsolved bug there.