Added methods for in-place Matrix row operations row_mult and row_add
These methods are added to the MutableRepMatrix class and so are inherited
by the Matrix class.
The common row operations of multiplying by a constant factor, adding a
constant multiple of one row to another, and swapping two rows are now
available as individual methods. Previously only the last of these was
available, along with a very general method row_op, which
is not intuitive to apply.

Examples
        ========

        >>> from sympy import eye
        >>> M = eye(3)
        >>> M.row_mult(1,7); M
        Matrix([
        [1, 0, 0],
        [0, 7, 0],
        [0, 0, 1]])

        >>> M = eye(3)
        >>> M.row_add(0, 2,3); M
        Matrix([
        [1, 0, 0],
        [0, 1, 0],
        [3, 0, 1]])