Rewrote the multiplication node to be simply a call to `sympy.Mul` without any advanced logic.
Moved the complexity of the `mul` node to the `adjacent_expressions` node.

Added support for OuterProduct in the LaTeX parser.

Refactored the differential variable detection code into its own function.

Renamed `differential_symbol_index` and `differential_symbol` to `differential_variable_index` and `differential_variable` because that's what the variables store.