Move GF(p) to_int method to domain rather than element
The to_int and __int__ methods of GF(p) are the only methods that
depend on the symmetric property of the domain. Having this dependence
at the element level rather than the domain level makes it not possible
to swap out the implementing of GF(p) for a more efficient
implementation that does not provide the to_int method or the symmetric
property.

The only place that the to_int method was being used directly was in
crypto. It is unclear if the crypto code actually needs the symmetric
property because all of the tested examples use GF(2) for which it has
no effect. The crypto code does not provide for the user to pass in a
domain so there is no way to use K.to_int(a) instead of a.to_int() in
order to preserve the existing behaviour.