besselk: add series expansion for non-integer order and fix integer order
NotImplementedError is now raised for unhandled cases (i.e. when the
order is not real.

We now also also return the correct Order when the order (n) of the
argument of the Bessel function is higher than the order requested by
the user.

For integer order:
* fix various edge cases and order
of returned term.
* Also remove unused branch of if-else: (k-nu)*k can never be zero for k in range(1,nu)

Re. the order of the returned term:
Since the O() term in a is multiplied by log(x), it is not the same as
the ones which are present (but removed in previous lines) in b and c.
This can sometimes lead to the wrong order of the truncated terms being
indicated in the final expression.