Fast `_perfect_power` implemented
`perfect_power` is slow(see #25138). So we implemented `_perfect_power`
which has less functionality. The arguments `candidates`, `big`, and
`factor` of `perfect_power` are absent. This speeds up the calculation of
perfect power. Use `_perfect_power` when
`candidates is None and big is True`. Otherwise, the previous logic is used.