Validate set inputs for linear_eq_to_matrix
Previous implementations allowed unordered sets as inputs for
linear_eq_to_matrix, leading to unpredictable outputs. This change
introduces input validation which rasie an error when sets are
accepted as input to ensure that the symbols are provided in a
predictable sequence, enforcing consistent behavior.

New test cases cover the scenarios where sets are passed to the
function, ensuring proper erroe responses.

Resolves: #25423