Make Float.is_integer always return None
This is a fix for #23731. Mostly just trying to see what consequences this
change has. It fixes some wrong results, for example:

>>> k = Symbol('k', integer=True)
>>> Eq(k/2, 0.5)
False

(now it is correctly unevaluated).

But I don't know yet if this breaks things.

Another possibility would be to return True if the Float is an integer that is
exactly represented, False if it has a nonzero fractional part, and None if it
is larger in magnitude than 2**prec.