calculus: extend singularities() for trig/hyp/inverses
Implement singularities searching for all 24 trigonometric and
hyperbolic functions and their inverses.
Tests are added.

Of particular note is that in some cases removable singularities are in
fact implicitely removed by cancelation:

In [3]: singularities(tanh(x)*cosh(x), x)
Out[3]: EmptySet

This is consistent with the regular trig functions, but see #10702 for a
discussion on this topic.