nonlinsolve: fix handling multiple ImageSets
Before this commit the original_imageset dict wasn't cleared for each
"result" the main for loop. This led to spilling of previous ImageSets
into later unrelated solutions.

In particular for the test for issue 16618 it led to "solutions" of the
form (2*n*pi+pi, 2*n*pi).

Also: the previous test answer was wrong.