solveset: better domain handling in invert()
This commit adds an intersection with the passed `domain` in more cases.
In particular `ImageSets` are now interesected with an `Interval` passed as
`domain`.

This commit also fixes the return value when EmptySet is passed to
invert_real or invert_complex. We now return the original symbol as
first entry in the tuple, signalling that the returned set is indeed the
solution in terms of the original variable.