nonlinsolve: fix ImageSet handling
Before this commit an ImageSet found as partial solution for one of the
uknowns led unconditionally to the creation of a new solution ImageSet
after its substitution into the next equation. But this substitution
could in fact lead to a simplification that reduced the infinite number
of solutions to a finite number of solutions. `_append_new_soln` then
encapsulated these solutions in a superfluous ImageSet which then
auto-evaluated to a simple FiniteSet which in turn triggered an assert()
marked as "internal error".

This commit fixes this issue by specifically checking for such
simplifications that completely remove the previous ImageSet's dummy
variable from the equation.

Unfortunately this is still not sufficient to make the test for
issue #17933 pass. A renaming of the unknowns works though.
(Which obviously raises questions.)