LaTeX parsing: add matrix, literal complex numbers (lark)
Overview of features added or modified:
1. Matrix parsing is added to the Lark-based parser.
2. Literal complex numbers are introduced via the LaTeX macro
\sympyi in order to facilitate working with matrices with complex
number entries.
3. Since the single quote (prime) symbol is sometimes used to denote
matrix transpose, that is introduced, and as a side-effect, primed
symbols are added to the Lark parser. The behavior of primes
on symbols is completely faithful to the input. For example, the input
"a''_{b'''}''''" generates sympy.Symbol("a''_{b'''}''''").

List of features added:
1. Imaginary unit: \sympyi
2. Matrix parsing (array and AMS-LaTeX matrix, pmatrix, bmatrix)
and detailed error handling
3. Determinant parsing (\det, vmatrix, |array matrix|,
|AMS-LaTeX matrix|)
4. Transpose via superscripts ^T, ^{T}, ^{'}, '. Applied multiple
times via '', ''', ..., ^{''}, ^{'''}, ...
5. Conjugate transpose via superscripts ^H, ^{H}, ^{*}. Applied
multiple times via ^{**}, ^{***}, ...
6. Matrix Trace: \sympytr
7. Matrix adjugate: \sympyadj
8. The above matrix functionality works on all matrix expressions
and not just individual literal matrices; for example, \det can
be applied to the sum of 2 LaTeX matrices.
9. Large number of new tests for all new functionality.

List of features modified:
1. Symbols (one letter, greek, multi-letter, subscripted) can have
primes (single quotes) on them in the Lark grammar.
2. In the Lark grammar, integral, derivative, and limit
expressions have been moved from _expression_mul to _expression_func
so that, for instance, integrals can be multiplied with each other.
3. In the Lark grammar, the rule adjacent_expressions has been
modified so that "fraction _expression" is now
"fraction _expression_mul". This is done to eliminate the incorrect
alternatives in the ambiguous interpretations of \frac{a}{b} x+y,
\frac{a}{b} + \frac{x}{y}, \frac{a}{b} - \frac{x}{y}, etc.
4. The Lark grammar now understands "+ expression" as just
"expression".

Relevant issues:
1. https://github.com/sympy/sympy/issues/14007
2. https://github.com/sympy/sympy/issues/25676
3. https://github.com/sympy/sympy/issues/25366

Signed-off-by: Soumendra Ganguly <soumendraganguly@gmail.com>