Refactoring of `_special_diop_DN`
The calculation of `F` has been changed to list comprehension and
`factorint` is now used. Since prime factorization was being done by trial
factoring, a speedup can be expected when `N` is large. Also, `F` was
changed from list to dict. Obviously `N // f**2` will not overlap and there
is no need to do a linear search.
The variable `i` has been removed; we only need to make a conditional
decision on break once every two times, so we use a for statement.