prover: Fixed _process_rule() function fixme issue, added explanations for process_rule() function
I provided additional information about the purpose and functionality of the process_rule method. I also modified the '_process_rule' method to ensure that the transformation is only applied when both 'b' and 'c' are not empty, addressing the issue raised by the FIXME comment '# (?) FIXME this is only correct when b & c != null !'.