Control: function addition for the phase criterion
I added a function to the TransferFunction class
called phase_criterion(z) and has an input.
This input can be a number that belongs to C and
if it belongs to the Root Locus of the transfer
function it returns True otherwise False. I think
it's a good start for creating extensions over the
Root Locus.

Example:

>>> system2 = TransferFunction((1), ((s**2+100*s+2600)*(s+25)*s), s)
>>> z1 = 37.1+65.2*I
>>> system2.phase_criterion(z1)
True

modified:   sympy/physics/control/lti.py
modified:   sympy/physics/control/tests/test_lti.py