Use dot, cross, outer as primary methods in Vector and Dyadic
Originally Vector and Dyadic were designed to use dot = __and__, cross =
__xor__, and outer = __or__ as well as the right side definitions of
these builtin operators. We have stopped advertising the use of these
methods in the documentation for users and this changes the Vector and
Dyadic classes so that we could more easily deprecate them in the
future. It also removes the use of the builtin operators within the code
of the two classes.