combinatorics: improved the FpGroups.index() method
Added the use of Lagrange's theorem to FpGroup.index() method so that it
returns |G|/|H| when G has finite order, by using FpGroup.subgroup() method.
I think there is an issue with this method, if the FpGroup has infinite order,
but is not picked up by FpGroup._is_infinite() method.
Added myself to .mailmap as I'm new contributor.