Fix cse treatment in lambdify with Derivatives
Before, when there were Derivatives in expr ans args given to lambdify
with cse enabled, there was an error because the cse treatment changed
the arguments of the Derivative object.

With this implementation, the expression is pre-treated by a function to
mask the instances of Derivative objects, then the cse process is
applied and finally we do a post-treatment to put back the Derivative
expressions in expr.