documentation: Clarify method to add name to .mailmap file
Clarified how to use two entries on the line in the .mailmap file at the
top of the "Adding your name and email address to .mailmap file." section.
Fixed references and links to "Configure git settings" which is now in
dev-setup.md file. Added the link reference to the section in dev-setup.md.