Add dependabot for updating GitHub Actions workflow versions
Note that this only enables dependabot for GitHub Actions workflow versions
(e.g., things like `actions/checkout@v3`). This does not enable it for things
like Python packages. We could do that later, but that's a separate
discussion.

This will run dependabot one a week (I believe on Mondays). The updates are
grouped, so there should only be at most one PR.

I have also added a workflow that enables auto-merge on these PRs, so that if
the checks pass, they will merge automatically. We can remove this if you want
to merge them manually, but I've found that the actions updates rarely break
anything so this just makes it so you can ignore dependabot.