diophantine: add support of cornacchia when a+b>m
Occasionally diophantine() and diop_quadratic() results in a DN form
$X^2 - DY^2 = N$ which cannot be handled by the current cornacchia()
properly. For example, previously

>>> eq1 = x**2 - 15*x + y**2 - 8*y # also mentioned in #18628
>>> diophantine(eq1)
{(0, 0), (15, 0), (15, 8), (0, 8)}
>>> eq1.subs({x: 16, y: 4})
0

>>> eq2 = 2*x**2 - 9*x + 4*y**2 - 8*y + 14
>>> diophantine(eq2)
set()
>>> eq2.subs({x: 2, y: 1})
0

Both the examples leads to a function call cornacchia(1, X, 1) with
X > 1, which turns out to have no solution for cornacchia(). The
solver can only be applied when $a+b<=m$,
ref: http://www.numbertheory.org/php/cornacchia.html
An additional check has been added before the algorithm based on the
fact that $ax^2 + by*2 = m$ admits a solution for $a + b > m$ only if
$x = 0$ or $y = 0$. A test function includes both the two examples
has been appended as well.