Better handle roots and cleaned projection facto
Had issues with convergence of root calculations. Solved it by adding factoring before the root computations as the convergence issue was just due to large coefficients. Also cleaned up the Hong projection set by removing constants and removing polynomials that are a sign away from another one. I think we can possibly remove those that are a constant away, but not sure!