Revised conditions for `erf.is_real`
We had not considered the possibility that `erf` and `erfc` could yield real
numbers when given complex numbers as input. Therefore, conditions such as
`is_real` and `is_extended_positive` now return True (or False) only when
the arguments are extended real numbers. In this context, `erf(I)` is not a
real number, but since it can no longer be determined here, it has been
removed from the test cases.

Additionally, the conditions for `erfcinv.is_infinite` have also been revised.