fix(polys): Only use Flint for GF(p) if p is prime
python-flint often crashes for operations with nmod when the modulus is
not prime. It is better to raise an exception insteaed. Performance of
GF(p) for non-prime p is not so important (SymPy itself never uses
non-prime p) so for simplicity we don't use Flint if the modulus is not
prime.

Ultimately it would probably be best if GF(p) for non prime p simply
raised an exception when the domain is created.