Add two friction classes `CoulombKineticFriction`, `CoulombKineticFriction2`, and their test classes `TestCoulombKineticFriction`, `CoulombKineticFriction2`, and `TestCylinderFriction`
`CoulombKineticFriction` is the original friction actuator that assumes the normal force is a positive scalar. I explained in the docstring which systems work well with this actuator and its limitations so that users can better understand. Also, the variable names are changed from `normal_force` and `viscous_coefficient` to `f_n` and `sigma` for consistency.

`CoulombKineticFriction2` is the new friction actuator that assumes the normal force is a vector. It does not include a viscous term since the calculated force should be either a scalar or vector. Variable names are also changed for consistency. The `to_loads` method is added to ensure compatibility with various pathways, including `WrappingPathway`. There is a problem with the force expression.

In `TestCoulombKineticFriction`, only the variable names are changed.

`TestCoulombKineticFriction2` is a new test class for `CoulombKineticFriction2`. The tested system is the same as in `TestCoulombKineticFriction`. There is a problem with force expression.

`TestCylinderFriction` is a new test class for `CoulombKineticFriction2`. It aims to test a cylinder sliding along a curve in 3D space. Due to the issue with the force expression in `CoulombKineticFriction2` and `TestCoulombKineticFriction2`, this class is incomplete.