Delete CoulombKineticFriction2 and its test classes and update `CoulombKineticFriction2`
Update the docstring in `CoulombKineticFriction` with example codes, and relax the explanation. Implement `tanh(self.s_t * v)` instead of `sign(v)`, where `s_t` is a constant for smooth transition. Modify the force expression to ensure that the force is zero when the normal force is zero, and standardise variable names based on a single paper (Rogner2017), except for `s_t`.