integrals: Update Laplace transform tests
Previously Mul.as_base_exp allowed the following:

  >>> (sqrt(x)*sqrt(y)).as_base_exp()
  (x*y, S.Half)

This implicitly assumes that sqrt(x)*sqrt(y) == sqrt(x*y) which is not
true for all possible values of x and y e.g. if x = y = -1 then we have

  sqrt(-1)*sqrt(-1) = I*I = -1

However

  sqrt((-1)*(-1)) = sqrt(1) = 1

This change in as_base_exp means that lots of Laplace transform results
come out differently. This commit updates the tests to reflect this. I
have not verified that these changes are correct but the change to
as_base_exp should prevent as_base_exp itself from returning incorrect
results and the changes here are the effect that it has had on the
output from laplace_transform().