rootof: Finding roots of polynomials with real algebraic coefficients
Previously, SymPy could not solve for the roots of polynomials with real
algebraic coefficients, e.g. radicals or root objects. This change computes
the "rational lift" of the polynomial, i.e., a polynomial with rational
coefficients whose roots are a superset of the original polynomial's, and
then numerically filters them.

Example:

Poly(x**3 + sqrt(2)*x**2 - 1, x).real_roots() == [
        rootof(x**6 - 2*x**4 - 2*x**3 + 1, 0)
]

Previously, this would fail.

It uses the new functionality in #26813 and #26816 for constructing the
extension and correct root counting. As well, it adapts code from #22943.