rootoftools: Modify when algebraic extension is constructed
In the previous commits, when solving for roots of polynomial
with algebraic coeffs, both rational and algebraic coeffs were
handled in the function _get_root(). Now, _get_root() checks
for the domain and then dispatches to the two new functions
_get_root_qq() or _get_root_alg(). The former is basically the
same as the master branch, the latter includes the lifting and
numerical filtering procedures. As well, it now assumes that the
polynomial already has an algebraic extension field. This way,
whether or not the extension is constructed can be handled "higher
up" in the chain.

Lastly, in _numerically_filter_roots(), it checks whether the same
root has already been processed to not loop over them again.