polytools: New method to numerically filter roots
The ._numerically_filter_roots() method takes a set of candidate
roots that are a superset of f's roots and uses increasing
numerical precision to filter them out until we equal f's root
count, counted with multiplicity. This is necessary for getting roots
of a polynomial with algebraic coefficients from the roots of its
rational lift, and is right now only used for that purpose.