polytools: New method root_multiplicity()
The new Poly.root_multiplicity() method finds the multiplicity
of a root with numerical evaluation of its derivatives. This is
necessary for the which_roots() method to work properly for
real roots, as count_roots() returns the count of unique roots.
This meant that the loop in which_roots() would not terminate,
until an overflow error. This was previously breaking some
examples in real_roots(). Therefore, the loop condition was
changed, and multiplicity computed after.

Examples of errors before the change:

>>> Poly((x-1)**2, x).which_roots([1, 1])
OverflowError: cannot convert float infinity to integer

>>> Poly((x-sqrt(2))**2, x, extension=True).real_roots()
OverflowError: cannot convert float infinity to integer

In the last example, observe the roots of the lift:

>>> Poly((x-sqrt(2))**2, x, extension=True).lift().real_roots()
[-sqrt(2), -sqrt(2), sqrt(2), sqrt(2)]

With the change, this works fine:

>>> Poly((x-sqrt(2))**2, x, extension=True).real_roots()
[sqrt(2), sqrt(2)]