rootoftools: _get_roots_alg now uses square-free factorization
Implementation of _get_roots_alg() now first computes the square-
free factorization, finds the roots of each factor, and then puts
them together and sorts. This is to avoid issues with multiplicity
of roots, which could arise when calling Poly.which_roots(). Also,
Poly.which_roots() now returns unique elements, which is useful. Added
a new method to ComplexRootOf called _sort_roots() which sorts a list
of roots in the usual order: reals first, then complexes by the real
and imaginary part, with conjugates together. Also removed the method
Poly.root_multiplicity() as it is not needed anymore and was also
not a good implementation. Lastly, added tests for polynomials with
algebraic coefficients with multiple roots.