rootoftools: Changed _sort_roots() method
Previously, _sort_roots() assumed that complex roots appeared
in conjugate pairs. However, this is not true if the polynomial
in question is over a Gaussian domain (e.g. x-I). So, had to
change the way it is sorting the roots. Then, an issue arose
where some roots are not directly comparable, and SymPy
would throw a TypeError. Therefore, I now try to compare roots
symbolically if possible, but in the case of TypeError, fall back
on numerical evaluation. This is not ideal but I think it is the
best we can do without having information like root isolating
intervals.