rootoftools: get_roots() now handles Gaussian domains
get_roots() now checks for Gaussian domains ZZ_I or QQ_I and then
dispatches to get_roots_alg(). This enables finding roots of, e.g.,
x**2 + I. This also works for, e.g., x**2 + sqrt(2)*I. Added a
unit test for this, as well as added a little statement about this
under all_roots() in polytools. Note that this only works when
calling all_roots() -- an error will be thrown if real_roots() is
called on such polynomials as root-counting is not supported over
this domain (although in the examples outlined there are no real
roots at all).