polytools: Split which_roots into methods for real and all
Split Poly.which_roots() into Poly.which_real_roots() and
Poly.which_all_roots(). Added new documentation for them about
the requirement for the polynomial to be square free. Added new
tests to test edge cases where candidates is not a superset or when
the polynomial is not square free. Lastly, changed the calls in
CRootOf._get_roots_alg() to these two methods.