Second Quantization: Fixed TypeError in FermionState
The base class FermionState, which FermionStateBra/Ket inherits from, had a bug in its _negate_holes function.

In the case where symbolic states where used, the index chould not be compared to the fermi level.

Fixed by using _only_above_fermi and making sure it is a number before negating.