Refactor all function arguments with mutable defaults
Previously, many functions have arguments with mutable default values, most commonly {} and []. This is not a good coding style and could lead to hidden bugs if the function modifies the object. According to python documentation, the preferred way is to set the default value to None, then explicitly test for None at the beginning of the functions. See https://docs.python.org/3/reference/compound_stmts.html#function-definitions

While it's hard to say whether any of these functions had ever experienced problems as a result, I think it's better to refactor them to prevent any potential complications.