polytools: Added subresultant polynomials and coefficients
Added functions for subresultant polynomials and subresultant
coefficients. These use the subresultant PRS (`subresultants()`) already
in SymPy. The subresultant polynomials are computed from the subresultant
PRS according to a certain rule, described in the comments. As for the
subresultant coefficients, we simply take the coefficient of x^{i-1} of
the ith subresultant polynomial. These functions are necessary for other
tasks, including eventually for cylindrical algebraic decomposition.