polys: Check if extension is True when constructing polys
In _parallel_dict_from_expr_if_gens(), added a check if extension
is True, and if the factor is algebraic, then this is a valid factor.
Previously, this would give an error if, e.g., CRootOf objects were
given. Example: Poly(CRootOf(x**5+x**3-1,0)*x+1, x, extension=True)
would yield an "... contains an element of the set of generators"
error. Now, the polynomial is constructed properly.