polytools: Refactored subresultant polynomial code
Previously, subresultant_polys() initialized a list of 0s and then
assigned appropriate elements. This has now been changed to appending
or extending the list as needed. This happens by keeping track of the
degrees of the subresultant polys that are being computed versus the
current length of the list, and basically appends 0s until the
appropriate length. Also, I fixed an issue that was appearing with the
0th element of the subresultant polys, where it was being scaled by
an incorrect factor. It should be correct now, due to the observation
that the 0th element is always just the resultant. Also, there is no
need to store r_{i-1} (previously called "prev"), as only its degree
is needed.