Master <!-- Your title above should be a short description of what was changed. Do not include the issue number in the title. --> Enabling Bidirectional computation of residue #### References to other Issues or PRs <!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact format, e.g. "Fixes #1234" (see https://tinyurl.com/auto-closing for more information). Also, please write a comment on that issue linking back to this pull request once it is open. -->
Fixes #27156

#### Brief description of what is fixed or changed
This PR updates the residue function to utilize the modified nseries method for series expansion. The nseries method now captures series expansions in both positive and negative directions, allowing for accurate residue calculation. The residue is defined as the coefficient of
1/(𝑥−𝑥0) in the power series expansion around
𝑥=𝑥0.
#### Other comments
This change enhances the robustness of residue calculations, ensuring correct behavior in scenarios where series behavior differs in positive and negative directions.

#### Release Notes

series
Updated the residue function to utilize the modified nseries method for series expansion, allowing residue calculations from both positive and negative directions.
Improved handling of coefficients in series expansion to ensure accurate residue extraction.
<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->

<!-- END RELEASE NOTES -->