fix(core): don't extract Float from Pow of Add
Fixes gh-27195

Reverts gh-16762

This fixes e.g.

    In [4]: (x**2 + 10)**2
    Out[4]:
             2
    ⎛ 2     ⎞
    ⎝x  + 10⎠

    In [5]: (x**2 + 10.0)**2
    Out[5]:
                      2
          ⎛     2    ⎞
    100.0⋅⎝0.1⋅x  + 1⎠

Extracting a float like this was added in gh-16762 commit
c1e25a445f1c27829ea8cf0e04dbdd6d749834ce.

This commit reverts this change because there is no reason to do this
and in general it is better to preserve the basic form of the expression
if it cannot be substantively simplified by simple rules.