fix(core): restrict the 2-arg Mul to -1
Since 3e835b5 from gh-716 there is the following behaviour:

  >>> 2*(x + y)
  2*x + 2*y

Specifically if there is a Mul with 2 args and one arg is a Rational
and the other an Add then the Rational is distributed into the Add.
More generally if there is a Mul and the arguments collapse to a number
and an Add e.g. Mul(2, 3.0, x + y) then the coefficient is distributed
into the Add.

There are various problems with this behaviour such as the fact that it
breaks associativity (a*b)*c != a*(b*c). There have been previous
attempts to remove this e.g. gh-17242 but it is not easy because so much
code now depends on it.

This commit attempts a more limited change in which the 2-arg Mul is
mostly removed but kept for one important special case: if we have
Mul(-1, Add(...)) then the -1 will be distributed into the Add. This
special case is important because it arises when subtracting to Adds
e.g.:

  >>> x + y - (x + y)
  0

It could perhaps be possible to special case negation and subtraction
somehow so that negating the terms of an Add happens specifically in
subtraction rather than more generally in a Mul but here we just
preserve the 2-arg behaviour for -1 specifically as a more limited
change to existing behaviour. Keeping this particular special case of
the 2-arg Mul dramatically reduces the number of test failures seen in
the test suite comared to removing automatic distribution altogether.